# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A.M.E.L.I.A. — API de Triagem Hospitalar
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  mensagem?: string;
}

export interface SuccessResponse {
  sucesso: boolean;
  mensagem?: string;
}

export interface CadastroInput {
  nome: string;
  cpf: string;
  cartaoSus?: string;
  dataNascimento?: string;
  telefone?: string;
  senha: string;
}

export interface LoginInput {
  /** CPF ou Cartão SUS */
  identificador: string;
  senha: string;
}

export type UsuarioPapel = (typeof UsuarioPapel)[keyof typeof UsuarioPapel];

export const UsuarioPapel = {
  paciente: "paciente",
  admin: "admin",
} as const;

export interface Usuario {
  id: number;
  nome: string;
  cpf: string;
  cartaoSus?: string;
  dataNascimento?: string;
  telefone?: string;
  papel: UsuarioPapel;
  criadoEm: string;
}

export interface AuthResponse {
  usuario: Usuario;
  token?: string;
}

export interface TriagemInput {
  descricaoSintomas: string;
  /**
   * @minimum 0
   * @maximum 10
   */
  nivelDor?: number;
  temperatura?: number;
  frequenciaCardiaca?: number;
  temFebre?: boolean;
  dificuldadeRespirar?: boolean;
  dorNoPeito?: boolean;
  conscienciaAlterada?: boolean;
}

export type TicketPrioridade =
  (typeof TicketPrioridade)[keyof typeof TicketPrioridade];

export const TicketPrioridade = {
  urgent: "urgent",
  moderate: "moderate",
  light: "light",
} as const;

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  waiting: "waiting",
  called: "called",
  attended: "attended",
  cancelled: "cancelled",
} as const;

export interface Ticket {
  id: number;
  usuarioId: number;
  nomeUsuario?: string;
  codigo: string;
  prioridade: TicketPrioridade;
  descricaoSintomas: string;
  nivelDor?: number;
  temperatura?: number;
  frequenciaCardiaca?: number;
  temFebre?: boolean;
  dificuldadeRespirar?: boolean;
  dorNoPeito?: boolean;
  conscienciaAlterada?: boolean;
  predicaoML?: string;
  confiancaML?: number;
  correcaoAdmin?: string;
  status: TicketStatus;
  criadoEm: string;
  chamadoEm?: string;
}

export type AtualizarStatusInputStatus =
  (typeof AtualizarStatusInputStatus)[keyof typeof AtualizarStatusInputStatus];

export const AtualizarStatusInputStatus = {
  waiting: "waiting",
  called: "called",
  attended: "attended",
  cancelled: "cancelled",
} as const;

export interface AtualizarStatusInput {
  status: AtualizarStatusInputStatus;
}

export type CorrecaoInputCorrecao =
  (typeof CorrecaoInputCorrecao)[keyof typeof CorrecaoInputCorrecao];

export const CorrecaoInputCorrecao = {
  urgent: "urgent",
  moderate: "moderate",
  light: "light",
} as const;

export interface CorrecaoInput {
  correcao: CorrecaoInputCorrecao;
}

export interface FilaEspera {
  urgente: Ticket[];
  moderado: Ticket[];
  leve: Ticket[];
}

export interface ChamarResponse {
  ticket?: Ticket;
  encontrado: boolean;
}

export interface Estatisticas {
  total: number;
  aguardando: number;
  urgente: number;
  moderado: number;
  leve: number;
  atendidos: number;
  correcoesPendentes: number;
}

export interface MLRetreinarResponse {
  sucesso: boolean;
  amostrasUsadas?: number;
  acuracia?: number;
  mensagem?: string;
}

export interface MLLog {
  id: number;
  treinadoEm: string;
  amostrasUsadas: number;
  acuracia?: number;
  notas?: string;
}

export type ListarTicketsParams = {
  status?: ListarTicketsStatus;
  priority?: ListarTicketsPriority;
};

export type ListarTicketsStatus =
  (typeof ListarTicketsStatus)[keyof typeof ListarTicketsStatus];

export const ListarTicketsStatus = {
  waiting: "waiting",
  called: "called",
  attended: "attended",
  cancelled: "cancelled",
} as const;

export type ListarTicketsPriority =
  (typeof ListarTicketsPriority)[keyof typeof ListarTicketsPriority];

export const ListarTicketsPriority = {
  urgent: "urgent",
  moderate: "moderate",
  light: "light",
} as const;