# lib/api-zod/src/generated/types/atualizarStatusInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A.M.E.L.I.A. — API de Triagem Hospitalar
 * OpenAPI spec version: 0.1.0
 */
import type { AtualizarStatusInputStatus } from "./atualizarStatusInputStatus";

export interface AtualizarStatusInput {
  status: AtualizarStatusInputStatus;
}