# lib/api-zod/src/generated/types/mLLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A.M.E.L.I.A. — API de Triagem Hospitalar
 * OpenAPI spec version: 0.1.0
 */

export interface MLLog {
  id: number;
  treinadoEm: Date;
  amostrasUsadas: number;
  acuracia?: number;
  notas?: string;
}