# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A.M.E.L.I.A. — API de Triagem Hospitalar
 * OpenAPI spec version: 0.1.0
 */
import type { TicketPrioridade } from "./ticketPrioridade";
import type { TicketStatus } from "./ticketStatus";

export interface Ticket {
  id: number;
  usuarioId: number;
  nomeUsuario?: string;
  codigo: string;
  prioridade: TicketPrioridade;
  descricaoSintomas: string;
  nivelDor?: number;
  temperatura?: number;
  frequenciaCardiaca?: number;
  temFebre?: boolean;
  dificuldadeRespirar?: boolean;
  dorNoPeito?: boolean;
  conscienciaAlterada?: boolean;
  predicaoML?: string;
  confiancaML?: number;
  correcaoAdmin?: string;
  status: TicketStatus;
  criadoEm: Date;
  chamadoEm?: Date;
}